fix: StopNoBuffer has consequent
StopNoBuffer means the buffer will be cleared out after the status is
sent. Therefore, we need to store the bytes in the context, and after
buffer is fully loaded then we need to set it back to the buffer
instance.
The fully loaded buffer indicated by endStream with value of true.

Signed-off-by: Ardika Bagus <me@ardikabs.com>